ProjetoFinal/A/requirements.txt
quoters
kafka-python
findspark